Updating language translating to Czech. More changes for sensors input to OSPy. For this testing use FW:1.02 and ESP32 board in folder: hardware_pcb/sensors_pcb_fw/. More infos with images is here: https://pihrt.com/elektronika/439-ospy-jak-pridat-a-pripojit-snimace. Sensor ESP32 FW: 1.02: Add AP manager for settings sensor from phone, tablet (4 minutes timeout). For run AP manager: push button least > 2 seconds. Password for AP Wi-Fi is "ospy-sensor-esp32", next open browser and type IP "192.168.1.1". Add saving to eeprom memory. Fast blinking LED if AP manager. Slow blinking LED if normal run. Add control relay from stations (plugin CLI Control in plugin manager): ex command: http://IP/securecode?re=1 (or re=0). In AP manager is active button  for control relay ON or OFF. Available only on the same network as OSPy. Now exists parameter "multi sensor" in sensors (for measure temperature 4xDS18B20 and Dry Contact, Leak Detector, Moisture, Motion) just one ESP32 board and all parameters can be sent and responded to in OSPy.

+++ ospy/webpages.py
@@ -163,6 +163,15 @@ class sensors_page(ProtectedPage):
 
         delete_all = get_input(qdict, 'delete_all', False, lambda x: True)
         search = get_input(qdict, 'search', False, lambda x: True) 
+        clean_all = get_input(qdict, 'clean_all', False, lambda x: True)
+
+        if clean_all:
+            for i in range(0, len(sensorSearch)):
+                try:
+                    del sensorSearch[int(i)]            
+                except:
+                    pass    
+            return self.core_render.sensors_search()
           
         if delete_all:
             while sensors.count() > 0:
@@ -241,6 +250,9 @@ class sensor_page(ProtectedPage):
             if 'com_type' in qdict:
                 sensor.com_type = int(qdict['com_type'])
 
+            if 'multi_type' in qdict:
+                sensor.multi_type = int(qdict['multi_type'])
+
             if 'log_samples' in qdict and qdict['log_samples'] == 'on':
                 sensor.log_samples = 1
             else:                                  
@@ -257,13 +269,13 @@ class sensor_page(ProtectedPage):
                 sensor.send_email = 0 
 
             if 'sample_rate_min' in qdict and 'sample_rate_sec' in qdict:
-                sensor.sample_rate = int(qdict['sample_rate_min'])*60 + int(qdict['sample_rate_sec'])                
+                sensor.sample_rate = int(qdict['sample_rate_min'])*60 + int(qdict['sample_rate_sec'])
 
             if 'sensitivity' in qdict:
                 sensor.sensitivity = int(qdict['sensitivity'])
 
             if 'stabilization_time_min' in qdict and 'stabilization_time_sec' in qdict:
-                sensor.stabilization_time = int(qdict['stabilization_time_min'])*60 + int(qdict['stabilization_time_sec'])                 
+                sensor.stabilization_time = int(qdict['stabilization_time_min'])*60 + int(qdict['stabilization_time_sec'])
 
             if 'trigger_low_threshold' in qdict:
                 sensor.trigger_low_threshold = qdict['trigger_low_threshold']
